misc/HtmlTestSuite/results/test-0539.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Grid Layout Test: grid sets longhands</title>
<link href="https://drafts.csswg.org/css-grid/#grid-shorthand" rel="help"/>
<meta content="grid supports the full grammar '&lt;'grid-template'&gt; | &lt;'grid-template-rows'&gt; / [ auto-flow &amp;&amp; dense? ] &lt;'grid-auto-columns'&gt;? | [ auto-flow &amp;&amp; dense? ] &lt;'grid-auto-rows'&gt;? / &lt;'grid-template-columns'&gt;'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/shorthand-testcommon.js"/>
</head>
<body>
<script>
// &lt;grid-template&gt;
test_shorthand_value('grid', 'none', {
  'grid-template-rows': 'none',
  'grid-template-columns': 'none',
  'grid-template-areas': 'none',

  'grid-auto-rows': 'auto',
  'grid-auto-columns': 'auto',
  'grid-auto-flow': 'row'
});

test_shorthand_value('grid', '10px / 20%', {
  'grid-template-rows': '10px',
  'grid-template-columns': '20%',
  'grid-template-areas': 'none',

  'grid-auto-rows': 'auto',
  'grid-auto-columns': 'auto',
  'grid-auto-flow': 'row'
});

// This could theoretically be serialized as "auto-flow / 10px",
// but spec mandates the 'grid-template-*' form when the
// 'grid-auto-*' properties are all initial.
test_shorthand_value('grid', 'none / 10px', {
  'grid-template-rows': 'none',
  'grid-template-columns': '10px',
  'grid-template-areas': 'none',

  'grid-auto-rows': 'auto',
  'grid-auto-columns': 'auto',
  'grid-auto-flow': 'row'
});

test_shorthand_value('grid', 'fit-content(calc(-0.5em + 10px)) / fit-content(calc(0.5em + 10px))', {
  'grid-template-rows': 'fit-content(calc(-0.5em + 10px))',
  'grid-template-columns': 'fit-content(calc(0.5em + 10px))',
  'grid-template-areas': 'none',

  'grid-auto-rows': 'auto',
  'grid-auto-columns': 'auto',
  'grid-auto-flow': 'row'
});

test_shorthand_value('grid',
                     '[header-top] "a a a"     [header-bottom]' +
                     '  [main-top] "b b b" 1fr [main-bottom]' +
                     '           / auto 1fr auto', {
  'grid-template-rows': '[header-top] auto [header-bottom main-top] 1fr [main-bottom]',
  'grid-template-columns': 'auto 1fr auto',
  'grid-template-areas': '"a a a" "b b b"',

  'grid-auto-rows': 'auto',
  'grid-auto-columns': 'auto',
  'grid-auto-flow': 'row'
});

test_shorthand_value('grid',
                     '  "a a a"' +
                     '  "b b b" 1fr' +
                     '/ auto 1fr auto', {
  'grid-template-rows': 'auto 1fr',
  'grid-template-columns': 'auto 1fr auto',
  'grid-template-areas': '"a a a" "b b b"',

  'grid-auto-rows': 'auto',
  'grid-auto-columns': 'auto',
  'grid-auto-flow': 'row'
});

test_shorthand_value('grid',
                     ' [] "a a a"     []' +
                     ' [] "b b b" 1fr []' +
                     '  / [] auto 1fr [] auto []', {
  'grid-template-rows': 'auto 1fr',
  'grid-template-columns': 'auto 1fr auto',
  'grid-template-areas': '"a a a" "b b b"',

  'grid-auto-rows': 'auto',
  'grid-auto-columns': 'auto',
  'grid-auto-flow': 'row'
});

// &lt;'grid-template-rows'&gt; / [ auto-flow &amp;&amp; dense? ] &lt;'grid-auto-columns'&gt;?
test_shorthand_value('grid',
                     '10px' +
                     ' / auto-flow dense' +
                     ' 20px', {
  'grid-template-rows': '10px',
  'grid-template-columns': 'none',
  'grid-template-areas': 'none',

  'grid-auto-rows': 'auto',
  'grid-auto-columns': '20px',
  'grid-auto-flow': 'column dense'
});

// [ auto-flow &amp;&amp; dense? ] &lt;'grid-auto-rows'&gt;? / &lt;'grid-template-columns'&gt;
test_shorthand_value('grid',
                     'auto-flow dense' +
                     ' 30px /' +
                     ' 40px', {
  'grid-template-rows': 'none',
  'grid-template-columns': '40px',
  'grid-template-areas': 'none',

  'grid-auto-rows': '30px',
  'grid-auto-columns': 'auto',
  'grid-auto-flow': 'dense'
});
</script>
</body>
</html>
